refactor: move prompt data conditioning adapters
