+++ tests/test_registry_scanner_safety.py
@@ -40,6 +40,7 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/nodes/image_nodes.py",
     "easyuse_anima/nodes/lora_nodes.py",
     "easyuse_anima/nodes/naia_nodes.py",
+    "easyuse_anima/nodes/prompt_data_nodes.py",
     "easyuse_anima/nodes/prompt_nodes.py",
     "easyuse_anima/nodes/wildcard_nodes.py",
     "easyuse_anima/profiles/__init__.py",
@@ -47,6 +48,9 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/profiles/mutation.py",
     "easyuse_anima/prompt/__init__.py",
     "easyuse_anima/prompt/correction.py",
+    "easyuse_anima/prompt/artist_mix.py",
+    "easyuse_anima/prompt/conditioning.py",
+    "easyuse_anima/prompt/data.py",
     "easyuse_anima/prompt/fields.py",
 }
 STATIC_IMPORT_FROM_RE = re.compile(
@@ -154,9 +158,13 @@ class RegistryScannerSafetyTests(unittest.TestCase):
             "easyuse_anima/nodes/image_nodes.py",
             "easyuse_anima/nodes/lora_nodes.py",
             "easyuse_anima/nodes/naia_nodes.py",
+            "easyuse_anima/nodes/prompt_data_nodes.py",
             "easyuse_anima/nodes/prompt_nodes.py",
             "easyuse_anima/nodes/wildcard_nodes.py",
             "easyuse_anima/prompt/correction.py",
+            "easyuse_anima/prompt/artist_mix.py",
+            "easyuse_anima/prompt/conditioning.py",
+            "easyuse_anima/prompt/data.py",
             "easyuse_anima/prompt/fields.py",
             "nodes.py",
             "prompt_translation.py",
